# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MonkeyCode API",
        "title": "MonkeyCode API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/workspace/files": {
            "get": {
                "description": "分页获取工作区文件列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "获取工作区文件列表",
                "operationId": "list-workspace-files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "编程语言筛选",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "下一页标识",
                        "name": "next_token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "分页",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词（文件路径）",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页多少条记录",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "工作区ID",
                        "name": "workspace_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.ListWorkspaceFileResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "创建一个新的工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "创建工作区文件",
                "operationId": "create-workspace-file",
                "parameters": [
                    {
                        "description": "文件信息",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateWorkspaceFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WorkspaceFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/workspace/files/batch": {
            "put": {
                "description": "批量更新多个工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "批量更新工作区文件",
                "operationId": "batch-update-workspace-files",
                "parameters": [
                    {
                        "description": "文件列表",
                        "name": "files",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BatchUpdateWorkspaceFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.WorkspaceFile"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "批量创建多个工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "批量创建工作区文件",
                "operationId": "batch-create-workspace-files",
                "parameters": [
                    {
                        "description": "文件列表",
                        "name": "files",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.BatchCreateWorkspaceFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.WorkspaceFile"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/workspace/files/by-path": {
            "get": {
                "description": "根据用户ID、项目ID和文件路径获取工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "根据路径获取工作区文件",
                "operationId": "get-workspace-file-by-path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "项目ID",
                        "name": "project_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件路径",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WorkspaceFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/workspace/files/sync": {
            "post": {
                "description": "同步本地文件到工作区，智能检测新增、修改和删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "同步工作区文件",
                "operationId": "sync-workspace-files",
                "parameters": [
                    {
                        "description": "同步信息",
                        "name": "sync",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.SyncWorkspaceFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.SyncWorkspaceFileResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/workspace/files/{id}": {
            "get": {
                "description": "根据文件ID获取工作区文件详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "根据ID获取工作区文件",
                "operationId": "get-workspace-file-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WorkspaceFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "更新指定的工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "更新工作区文件",
                "operationId": "update-workspace-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "file",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateWorkspaceFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.WorkspaceFile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的工作区文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WorkspaceFile"
                ],
                "summary": "删除工作区文件",
                "operationId": "delete-workspace-file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.BatchCreateWorkspaceFileReq": {
            "type": "object",
            "required": [
                "files",
                "user_id",
                "workspace_id"
            ],
            "properties": {
                "files": {
                    "description": "文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CreateWorkspaceFileReq"
                    }
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "workspace_id": {
                    "description": "工作区ID",
                    "type": "string"
                }
            }
        },
        "domain.BatchUpdateWorkspaceFileReq": {
            "type": "object",
            "required": [
                "files"
            ],
            "properties": {
                "files": {
                    "description": "文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.UpdateWorkspaceFileReq"
                    }
                }
            }
        },
...
        "domain.CreateWorkspaceFileReq": {
            "type": "object",
            "required": [
                "hash",
                "path",
                "user_id",
                "workspace_id"
            ],
            "properties": {
                "content": {
                    "description": "文件内容",
                    "type": "string"
                },
                "hash": {
                    "description": "文件哈希",
                    "type": "string"
                },
                "language": {
                    "description": "编程语言",
                    "type": "string"
                },
                "path": {
                    "description": "文件路径",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小",
                    "type": "integer"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "workspace_id": {
                    "description": "工作区ID",
                    "type": "string"
                }
            }
        },
...
        "domain.ListWorkspaceFileResp": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceFile"
                    }
                },
                "has_next_page": {
                    "type": "boolean"
                },
                "next_token": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SyncWorkspaceFileReq": {
            "type": "object",
            "required": [
                "files",
                "user_id",
                "workspace_id"
            ],
            "properties": {
                "files": {
                    "description": "要同步的文件列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.CreateWorkspaceFileReq"
                    }
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "workspace_id": {
                    "description": "工作区ID",
                    "type": "string"
                }
            }
        },
        "domain.SyncWorkspaceFileResp": {
            "type": "object",
            "properties": {
                "created": {
                    "description": "新创建的文件",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceFile"
                    }
                },
                "deleted": {
                    "description": "删除的文件ID",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "description": "处理的文件总数",
                    "type": "integer"
                },
                "updated": {
                    "description": "更新的文件",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WorkspaceFile"
                    }
                }
            }
        },
...
        "domain.UpdateWorkspaceFileReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "content": {
                    "description": "文件内容",
                    "type": "string"
                },
                "hash": {
                    "description": "文件哈希",
                    "type": "string"
                },
                "id": {
                    "description": "文件ID",
                    "type": "string"
                },
                "language": {
                    "description": "编程语言",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小",
                    "type": "integer"
                }
            }
        },
...
        "domain.WorkspaceFile": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "文件内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "integer"
                },
                "hash": {
                    "description": "文件哈希",
                    "type": "string"
                },
                "id": {
                    "description": "文件ID",
                    "type": "string"
                },
                "language": {
                    "description": "编程语言",
                    "type": "string"
                },
                "path": {
                    "description": "文件路径",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "integer"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                },
                "workspace_id": {
                    "description": "工作区ID",
                    "type": "string"
                }
            }
        },
        "web.Resp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}